refactor(ui): split alert page logic into hooks
