+++ openmetadata-ui/src/main/resources/ui/src/components/Alerts/AlertFormSourceItem/AlertFormSourceItem.interface.ts
@@ -11,8 +11,8 @@
  *  limitations under the License.
  */
 
-import { FilterResourceDescriptor } from '../../../generated/events/filterResourceDescriptor';
-
 export interface AlertFormSourceItemProps {
-  filterResources?: FilterResourceDescriptor[];
+  filterResources?: {
+    name?: string;
+  }[];
 }
