+++ openmetadata-ui/src/main/resources/ui/src/pages/AlertDetailsPage/AlertDetailsPage.interface.ts
@@ -11,6 +11,48 @@
  *  limitations under the License.
  */
 
+import type { TabsProps } from 'antd';
+import type { ReactNode } from 'react';
+import { AlertDetailTabs } from '../../enums/Alerts.enum';
+import {
+  EntityReference,
+  EventSubscription,
+} from '../../generated/events/eventSubscription';
+
 export interface AlertDetailsPageProps {
-  isNotificationAlert: boolean;
+  isNotificationAlert?: boolean;
 }
+
+export interface AlertDetailsPermissions {
+  deletePermission: boolean;
+  editDescriptionPermission: boolean;
+  editOwnersPermission: boolean;
+  editPermission: boolean;
+  viewPermission: boolean;
+}
+
+export interface UseAlertDetailsPageReturn extends AlertDetailsPermissions {
+  alertDetails?: EventSubscription;
+  alertIcon: ReactNode;
+  breadcrumb: {
+    name: string;
+    url: string;
+  }[];
+  extraInfo: ReactNode;
+  handleAlertDelete: () => Promise<void>;
+  handleAlertEdit: () => Promise<void>;
+  handleAlertSync: () => Promise<void>;
+  handleTabChange: (activeKey: string) => void;
+  hideDeleteModal: () => void;
+  isSyncing: boolean;
+  loadingCount: number;
+  onDescriptionUpdate: (description: string) => Promise<void>;
+  onOwnerUpdate: (owners?: EntityReference[]) => Promise<void>;
+  ownerLoading: boolean;
+  setShowDeleteModal: (show: boolean) => void;
+  showDeleteModal: boolean;
+  tab: AlertDetailTabs;
+  tabItems: TabsProps['items'];
+}
+
+export type AlertDetailsContentProps = UseAlertDetailsPageReturn;
